refactor(code-review-charts): create separate component for top 10 pull requests by comments
